refactor: separate TestBalances as utils
