+++ packages/stack/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./scripts/graphql/server";
 export * from "./presets/config";
 export * from "./presets/modules/types";
 export * from "./presets/modules/utils";
